+++ packages/app/src/cli/index.ts
@@ -6,6 +6,7 @@ import { Command } from 'commander';
 import { pkgRoot, DEFAULT_PORT } from './constants';
 import { registerStartCommand } from './commands/start';
 import { registerTuiCommand } from './commands/tui';
+import { registerRunCommand } from './commands/run';
 
 function getVersion(): string {
   try {
@@ -26,5 +27,6 @@ program
 
 registerStartCommand(program);
 registerTuiCommand(program);
+registerRunCommand(program);
 
 program.parse();